A-Assertions (#1)
* Add assert to ensure that storage is not null in TaskList

This assertion is intended to catch situations where the storage field
may be null. The assertion is added just before attempting to use the
storage field like in storage.saveTasks() or storage.loadTasks().

Given that Storage is a custom class, this additional check is useful to
catch potential bugs in my implementation.

* Add assert to ensure that taskList is not null in relevant Commands

Certain Commands contain a reference to taskList, and mutate the list.
Having the assert here helps to spot bugs if somehow the taskList is
null.